requirements.txt
dbt-core>=1.5.0
dbt-duckdb>=1.5.0
pandas>=1.3.0
jupyter>=1.0.0
